+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Added ability to select and export multiple maps from "Zones" at the same time.
 - Added new expansion icons for filtering + listfile views.
 - Added animation support in the 3D model viewer for models with external skeletons.
+- Added character animation previews in the "Characters" tab.
 - Reduced load time of characters tab from 50s to 219ms (230x faster).
 - Removed redundant 'Show M2', 'Show M3', and 'Show WMO' options.
 - Removed 'Show File Data IDs', this is now default behavior.